requirements.txt
streamlit
langchain
google-generativeai
python-dotenv
chromadb
PyMuPDF
